Move only_for_onprem decorator to api/utils.py
